+++ vrp-pragmatic/tests/features/clustering/capacity_vicinity_test.rs
@@ -163,7 +163,10 @@ fn can_vary_cluster_size_based_on_capacity_impl(
                     reasons: vec![UnassignedJobReason {
                         code: "CAPACITY_CONSTRAINT".to_string(),
                         description: "does not fit into any vehicle due to capacity".to_string(),
-                        detail: Some(UnassignedJobDetail { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 })
+                        details: Some(vec![UnassignedJobDetail {
+                            vehicle_id: "my_vehicle_1".to_string(),
+                            shift_index: 0
+                        }])
                     }]
                 })
                 .collect()),