+++ vrp-pragmatic/tests/features/compatibility/basic_compatibility.rs
@@ -75,7 +75,7 @@ fn can_unassign_job_due_to_compatibility() {
         vec![UnassignedJobReason {
             code: "COMPATIBILITY_CONSTRAINT".to_string(),
             description: "cannot be assigned due to compatibility constraint".to_string(),
-            detail: Some(UnassignedJobDetail { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 })
+            details: Some(vec![UnassignedJobDetail { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 }])
         }]
     );
 }