+++ vrp-pragmatic/tests/features/limits/max_distance.rs
@@ -36,7 +36,7 @@ fn can_limit_by_max_distance() {
                 reasons: vec![UnassignedJobReason {
                     code: "MAX_DISTANCE_CONSTRAINT".to_string(),
                     description: "cannot be assigned due to max distance constraint of vehicle".to_string(),
-                    detail: None
+                    details: None
                 }]
             }]),
             ..create_empty_solution()
@@ -76,7 +76,7 @@ fn can_handle_empty_route() {
                 reasons: vec![UnassignedJobReason {
                     code: "MAX_DISTANCE_CONSTRAINT".to_string(),
                     description: "cannot be assigned due to max distance constraint of vehicle".to_string(),
-                    detail: None,
+                    details: None,
                 }]
             }]),
             ..create_empty_solution()