+++ vrp-pragmatic/tests/features/limits/tour_size.rs
@@ -78,7 +78,7 @@ fn can_skip_job_from_multiple_because_of_tour_size() {
                 reasons: vec![UnassignedJobReason {
                     code: "TOUR_SIZE_CONSTRAINT".to_string(),
                     description: "cannot be assigned due to tour size constraint of vehicle".to_string(),
-                    detail: Some(UnassignedJobDetail { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 }),
+                    details: Some(vec![UnassignedJobDetail { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 }]),
                 }]
             }]),
             ..create_empty_solution()