+++ vrp-pragmatic/tests/features/priorities/basic_order.rs
@@ -153,7 +153,7 @@ fn can_follow_order_when_prioritized_property_set() {
             reasons: vec![UnassignedJobReason {
                 code: "MAX_DISTANCE_CONSTRAINT".to_string(),
                 description: "cannot be assigned due to max distance constraint of vehicle".to_string(),
-                detail: Some(UnassignedJobDetail { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 }),
+                details: Some(vec![UnassignedJobDetail { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 }]),
             }]
         }])
     );