+++ vrp-pragmatic/tests/features/priorities/basic_value.rs
@@ -44,7 +44,7 @@ fn can_prefer_jobs_with_more_value_impl(objectives: Option<Vec<Vec<Objective>>>)
             reasons: vec![UnassignedJobReason {
                 code: "CAPACITY_CONSTRAINT".to_string(),
                 description: "does not fit into any vehicle due to capacity".to_string(),
-                detail: Some(UnassignedJobDetail { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 })
+                details: Some(vec![UnassignedJobDetail { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 }])
             }]
         }
     );