+++ vrp-pragmatic/tests/helpers/solution.rs
@@ -108,14 +108,7 @@ pub fn create_empty_tour() -> Tour {
 }
 
 pub fn create_empty_solution() -> Solution {
-    Solution {
-        statistic: Default::default(),
-        tours: vec![],
-        unassigned: None,
-        violations: None,
-        resources: None,
-        extras: None,
-    }
+    Solution { statistic: Default::default(), tours: vec![], unassigned: None, violations: None, extras: None }
 }
 
 pub fn get_ids_from_tour(tour: &Tour) -> Vec<Vec<String>> {